Separate event store from event stream (#7592)
