+++ openhands/server/routes/feedback.py
@@ -2,8 +2,8 @@ from fastapi import APIRouter, Request, status
 from fastapi.responses import JSONResponse
 
 from openhands.core.logger import openhands_logger as logger
+from openhands.events.async_event_store_wrapper import AsyncEventStoreWrapper
 from openhands.events.serialization import event_to_dict
-from openhands.events.stream import AsyncEventStreamWrapper
 from openhands.server.data_models.feedback import FeedbackDataModel, store_feedback
 from openhands.utils.async_utils import call_sync_from_async
 
@@ -34,11 +34,11 @@ async def submit_feedback(request: Request, conversation_id: str) -> JSONRespons
     # Assuming the storage service is already configured in the backend
     # and there is a function to handle the storage.
     body = await request.json()
-    async_stream = AsyncEventStreamWrapper(
+    async_store = AsyncEventStoreWrapper(
         request.state.conversation.event_stream, filter_hidden=True
     )
     trajectory = []
-    async for event in async_stream:
+    async for event in async_store:
         trajectory.append(event_to_dict(event))
     feedback = FeedbackDataModel(
         email=body.get('email', ''),
