+++ openhands/server/routes/trajectory.py
@@ -2,8 +2,8 @@ from fastapi import APIRouter, Request, status
 from fastapi.responses import JSONResponse
 
 from openhands.core.logger import openhands_logger as logger
+from openhands.events.async_event_store_wrapper import AsyncEventStoreWrapper
 from openhands.events.serialization import event_to_trajectory
-from openhands.events.stream import AsyncEventStreamWrapper
 
 app = APIRouter(prefix='/api/conversations/{conversation_id}')
 
@@ -22,11 +22,11 @@ async def get_trajectory(request: Request) -> JSONResponse:
         events.
     """
     try:
-        async_stream = AsyncEventStreamWrapper(
+        async_store = AsyncEventStoreWrapper(
             request.state.conversation.event_stream, filter_hidden=True
         )
         trajectory = []
-        async for event in async_stream:
+        async for event in async_store:
             trajectory.append(event_to_trajectory(event))
         return JSONResponse(
             status_code=status.HTTP_200_OK, content={'trajectory': trajectory}
